Update DELETE blog so can only be deleted by user who added it

+++ part4/bloglist/controllers/blogs.js
@@ -12,7 +12,6 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response, next ) => {
     const body = request.body
-    console.log('request here', request.body)
 
     try {
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
@@ -43,9 +42,34 @@ blogsRouter.post('/', async (request, response, next ) => {
     }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
+blogsRouter.delete('/:id', async (request, response, next) => {
+    const body = request.body
+
+    try {
+    // Verify token validity and return corresponding user
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    console.log('What is this', decodedToken)
+    const user = await User.findById(decodedToken.id)
+    
+    // fetch the blog first, then check its parameters
+    const blog = await Blog.findById(request.params.id)
+    console.log(blog)
+
+    //compare parameters
+    if (blog.user.toString() === decodedToken.id.toString()) {
+        console.log("this is the correct user")
+        await Blog.findByIdAndDelete(request.params.id) 
+    } else {
+        return response.status(400).json({ error: 'Incorrect user, you do not have the rights to delete this blog' })
+    }
+
     response.status(204).end()
+    } catch (error) {
+        next (error)
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
